improve: cleanup entrypoint
